[FIX][I] Stop forwarding activities for IPN changes
This patch makes FileSystemChangeListener check if the file of an IntelliJ
event is shared before forwarding it. This allows to avoid forwarding
activities while unpacking the archive during the IPN, since completely
shared projects are registered as so only after the changes are actually
written to disk. Resources of a partially shared project are also added
individually after having been written to disk.

The filesystem implementation does not use the VFS yet so events for
file modifications will be fired asynchronously. Because of that this
patch will not fix the bug on the current master. This also means that
refreshing the local filesystem in the projectResourcesAvailable hook of
the SessionLifecycleListener, which is called after the resources all
shared projects are added to the session, is pretty much useless.

In other words, if you read this yours truly has submitted this patch
without checking it beforehand :)

Change-Id: Ic31e9302035c8e7eaacf35b671e429cdf593f947