[INTERNAL][I] Stop using workspace runnables in FileUtils
This patch focuses on removing the superfluous use of workspace runnables
in FileUtils, since the operations do not require to be atomic and the
IntelliJ implementation of IWorkspace#run do not do anything.

This patch is similar to changes made by commit 3adb19a [1] for the Eclipse version of
FileUtils.

[1] https://github.com/saros-project/saros/commit/3adb19a193a77b1e8515afc8a6113c3a1946dd15

Change-Id: I797296893593f475fc222d76fe8dc4fbad51dfa6