[INTERNAL][I] Correct use of IResource API
This patch fixes the way LocalEditorHandler retrieves the IResource for a
VirtualFile. It computes the project-relative path of the file and use
IProject#findMember to find the corresponding IResource only if it exists.

This fixes a bug where an incorrect resource would be built, only to throw
in the constructor of SPath, instead of returning null.

Change-Id: I47c28ea2d03e5a1dd1ae93081a9002670c5c1d0f